Day 2 File
Takes a comma-separated program string, parses it into a vector of ints, and then runs the associated program. 
Current Functionality: Basic addition and multiplication of two elements at given memory locations, stored to a third memory location. Halts when given halt command. Returns an error on bad input, operators, or being asked to alter a memory location that does not exist/is out of bounds.